requirements.txt
asgiref
boto3
botocore
certifi
chardet
Django
django-admin-thumbnails
django-session-timeout
django-storages
django-environ
idna
jmespath
Pillow
python-dateutil
python-decouple
pytz
requests
s3transfer
six
sqlparse
urllib3
django-rest-framework

